+++ src/pages/ReactHookForm/ControlledInput.tsx
@@ -1,58 +1,49 @@
+import { useState, useId } from 'react';
 import classNames from 'classnames';
-import { useId, useState, ChangeEvent } from 'react';
-import { FormItem, stringFormItem } from '../../types';
+import { UseFormRegister } from 'react-hook-form';
+import { FormItem } from '../../types';
 import ProgressPasswordStrength from '../../components/ProgressPasswordStrength';
 
 interface InputWrapperProps<K extends keyof FormItem> {
   name: K;
   label: string;
   type: 'text' | 'email' | 'password' | 'checkbox' | 'number' | 'select' | 'file';
-  invalidFeedback: stringFormItem;
+  register: UseFormRegister<FormItem>;
+  passwordForIndicator?: string;
+  errors?: string;
   options?: { value: string; label: string }[];
   datalistOptions?: string[];
-  onFileChange?: (file: FileList | undefined) => void;
 }
 
-const InputWrapper = <K extends keyof FormItem>({
+const ControlledInput = <K extends keyof FormItem>({
   name,
   label,
   type,
-  invalidFeedback,
+  register,
+  errors,
+  passwordForIndicator,
   options,
   datalistOptions,
-  onFileChange,
 }: InputWrapperProps<K>) => {
   const id = useId();
   const [showPassword, setShowPassword] = useState(false);
-  const [password, setPassword] = useState('');
 
   const togglePasswordVisibility = () => {
     setShowPassword((prevState) => !prevState);
-  };
-
-  const handleFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const files = event.target.files || undefined;
-    if (onFileChange) {
-      onFileChange(files);
-    }
-  };
-
-  const handlePasswordChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setPassword(event.target.value);
   };
 
   const classNameInput = classNames({
     'form-check-input': type === 'checkbox',
     'me-2': type === 'checkbox',
     'form-control': type !== 'checkbox',
-    'is-invalid': invalidFeedback[name].length > 0,
+    'is-invalid': errors,
   });
 
   let inputElement: JSX.Element;
 
   if (type === 'select' && options) {
     inputElement = (
-      <select className={classNameInput} id={id} name={name}>
+      <select className={classNameInput} id={id} {...register(name)}>
         <option value=''>Select an option</option>
         {options.map((option) => (
           <option key={option.value} value={option.value}>
@@ -64,24 +55,18 @@ const InputWrapper = <K extends keyof FormItem>({
   } else if (type === 'password') {
     inputElement = (
       <>
-        <input
-          type={showPassword ? 'text' : 'password'}
-          className={classNameInput}
-          id={id}
-          name={name}
-          onChange={handlePasswordChange}
-        />
+        <input type={showPassword ? 'text' : 'password'} className={classNameInput} id={id} {...register(name)} />
         <button type='button' onClick={togglePasswordVisibility} className='btn btn-outline-secondary ms-2'>
           {showPassword ? 'Hide' : 'Show'}
         </button>
       </>
     );
   } else if (type === 'file') {
-    inputElement = <input type='file' className={classNameInput} id={id} name={name} onChange={handleFileChange} />;
+    inputElement = <input type='file' className={classNameInput} id={id} {...register(name)} />;
   } else if (type === 'text' && datalistOptions) {
     inputElement = (
       <>
-        <input type='text' className={classNameInput} id={id} name={name} list={`${id}-datalist`} />
+        <input type='text' className={classNameInput} id={id} list={`${id}-datalist`} {...register(name)} />
         <datalist id={`${id}-datalist`}>
           {datalistOptions.map((option) => (
             <option key={option} value={option} />
@@ -92,15 +77,15 @@ const InputWrapper = <K extends keyof FormItem>({
   } else if (type === 'checkbox') {
     inputElement = (
       <div className='checkbox-wrapper'>
-        <input type='checkbox' className={classNameInput} id={id} name={name} />
+        <input type='checkbox' className={classNameInput} id={id} {...register(name)} />
         <label htmlFor={id} className='form-check-label'>
           {label}
         </label>
-        <div className='invalid-feedback'>{invalidFeedback[name]}</div>
+        <div className='invalid-feedback'>{errors}</div>
       </div>
     );
   } else {
-    inputElement = <input type={type} className={classNameInput} id={id} name={name} />;
+    inputElement = <input type={type} className={classNameInput} id={id} {...register(name)} />;
   }
 
   return (
@@ -110,13 +95,13 @@ const InputWrapper = <K extends keyof FormItem>({
           {label}
         </label>
       )}
-      {(name === 'password' || name === 'repeatPassword') && <ProgressPasswordStrength password={password} />}
+      {type === 'password' && <ProgressPasswordStrength password={passwordForIndicator ?? ''} />}
       <div className='input-group has-validation'>
         {inputElement}
-        <div className='invalid-feedback'>{invalidFeedback[name]}</div>
+        <div className='invalid-feedback'>{errors}</div>
       </div>
     </div>
   );
 };
 
-export default InputWrapper;
+export default ControlledInput;
